Move `schema/utils` to `components/Meditor` directory
