--- web/src/utils/typing.ts
@@ -1,3 +1,0 @@
-export const isRecord = (obj: unknown): obj is Record<string, unknown> => (
-  typeof obj === 'object' && obj !== null && !Array.isArray(obj)
-);
